+++ lib/loader.js
@@ -5,8 +5,7 @@
 const loaderUtils = require('loader-utils');
 
 const processCss = require('./processCss');
-const getImportPrefix = require('./getImportPrefix');
-const compileExports = require('./compile-exports');
+const { getImportPrefix, compileExports } = require('./utils');
 
 module.exports = function loader(content, map) {
   const callback = this.async();
