+++ lib/processCss.js
@@ -13,7 +13,7 @@ const modulesValues = require('postcss-modules-values');
 const cssLoaderParser = require('./postcss-css-loader-parser');
 
 const CssSyntaxError = require('./CssSyntaxError');
-const getLocalIdent = require('./getLocalIdent');
+const { getLocalIdent } = require('./utils');
 
 module.exports = function processCss(inputSource, inputMap, options, callback) {
   const { query } = options;
